Helper Functions

The following functions help in various task.

Index

DifferentialMobilityAnalyzers.clean

DifferentialMobilityAnalyzers.Σ

Functions

DifferentialMobilityAnalyzers.Σ — Method

Σ(f, i)

Defined as shorthand:

Σ(f, i) = mapreduce(f, +, 1:i)

The function evaluates f(X) for X = [1,...,i] and sums the result. If f(X) evaluates to a vector or SizeDistribution, the sum is the sum of the vectors or SizeDistributions.

Example Usage

Tc = getTc(Λ)
Σ(k -> Tc(k,100.0),2)  # evaluate the sum of Tc(1, 100.0), Tc(2, 100.0)

DifferentialMobilityAnalyzers.clean — Method

clean(x)

Defined as shorthand:

clean(x) = map(x -> x < 0.0 ? 0.0 : x, x)

The function removes negative numbers and set them zero. It is used to cleanup inverted size distribution data, which may contain small negative values from inversion noise.