Physics

The following functions are used to define the physics of the instrument.

Index

DifferentialMobilityAnalyzers.Tl

DifferentialMobilityAnalyzers.cc

DifferentialMobilityAnalyzers.dab

DifferentialMobilityAnalyzers.dtoz

DifferentialMobilityAnalyzers.getTc

DifferentialMobilityAnalyzers.vtoz

DifferentialMobilityAnalyzers.Ω

DifferentialMobilityAnalyzers.Ωav

DifferentialMobilityAnalyzers.η

Functions

DifferentialMobilityAnalyzers.η — Method

η(Λ::DMAconfig)

η is the viscosity of air in [Pa s] and depends on temperature [K]. Temperature is taken from the DMA configuration. Currently only dry air is supported.

$\eta = 1.83245\times10^{-5} \exp \left(1.5 \ln \left[\frac{T}{296.1}\right]\right)\left (\frac{406.55}{T+110.4} \right)$

Example Usage

t,p = 295.15, 1e5                        
qsa,qsh = 1.66e-5, 3.33e-6                     
r₁,r₂,l = 9.37e-3,1.961e-2,0.44369               
Λ = DMAconfig(t,p,qsa,qsh,r₁,r₂,l,0.0,:-,6,:cylindrical) 
viscosity = η(Λ)

DifferentialMobilityAnalyzers.cc — Method

cc(Λ::DMAconfig, d)

Cunningham slip-flow correction factor. The slip flow correction accounts for the decreased drag particles experience relative to Stokes' drag force when particle size approaches the scale of the mean free path of air. It is computed following Hinds (1999) Eq. 3.20. Temperature and pressure are taken from the DMA configuration. The units of diameter are in [m] and the function accepts scalars or arrays.

$c_c = 1+\frac{\lambda}{d_p} \left(2.34+1.05 \exp \left[-0.39 \frac{d_p}{\lambda}\right]\right)$

where $d_p$ is the particle diameter and $\lambda$ is the mean free path of air, which is computed as a function of pressure and temperature.

Example Usage

Dp = exp10.(range(log10(1e-9), stop=log10(1000e-9), length=100))
t,p = 295.15, 1e5                        
qsa,qsh = 1.66e-5, 3.33e-6                     
r₁,r₂,l = 9.37e-3,1.961e-2,0.44369               
Λ = DMAconfig(t,p,qsa,qsh,r₁,r₂,l,0.0,:-,6,:cylindrical) 
correction = cc(Λ, Dp)

DifferentialMobilityAnalyzers.dab — Method

dab(Λ::DMAconfig, d)

The diffusion coefficient of particles in air, $d_{ab}$, describes the random displacement of particles due to Brownian motion. It is computed via the Stokes-Einstein relation (Hinds, 1999, Eq. 7.20). Temperature and pressure are taken from the DMA configuration. The units of diameter are in [m] and the function accepts scalars or arrays.

$d_{ab} = \frac{k_bTc_c}{3\pi\eta d_p}$

where $k_b$ is Boltzmann's constant and $\eta$ is the viscosity of air in [Pa s], $c_c$ is the Cunningham slip flow correction and $d_p$ is the particle diameter. $d_{ab}$ is in [m² s⁻¹].

Example Usage

Dp = exp10.(range(log10(1e-9), stop=log10(1000e-9), length=100))
t,p = 295.15, 1e5                        
qsa,qsh = 1.66e-5, 3.33e-6                     
r₁,r₂,l = 9.37e-3,1.961e-2,0.44369               
Λ = DMAconfig(t,p,qsa,qsh,r₁,r₂,l,0.0,:-,6,:cylindrical) 
diffusion_coefficient = dab(Λ,Dp)

DifferentialMobilityAnalyzers.dtoz — Method

dtoz(Λ::DMAconfig, d)

The function returns the mobility $z$ according to

$d_p = \frac{kec_c}{3\pi \eta z^s}$

where $e$ is the elementary charge, $k$ is the number of charges on the particle, $c_c$ is the Cunningham correction factor, and $\eta$ is the viscosity of the fluid. The diameter in dtoz is in units of [m].

Example Usage

t,p = 295.15, 1e5                        
qsa,qsh = 1.66e-5, 3.33e-6                     
r₁,r₂,l = 9.37e-3,1.961e-2,0.44369               
Λ = DMAconfig(t,p,qsa,qsh,r₁,r₂,l,0.0,:-,6,:cylindrical) 
mobility = dtoz(Λ,dp*1e-9) # [m2 V-1 s-1]

DifferentialMobilityAnalyzers.vtoz — Method

vtoz(Λ::DMAconfig, v)

Converts between voltage and selected mobility.

For the cylindrical DMA and balanced flows:

$z^s = \frac{q_{sh}}{2\pi l v} \ln \left(\frac{r_2}{r_1}\right)$

For the radial DMA and balanced flows:

$z^s = \frac{q_{sh} l}{\pi v \left({r_2}^2 - {r_1}^2\right)}$

where $v$is the potential applied between the inner and out section of the annulus,r_1,r_2, andlare the dimensions of the cylindrical DMA andq_{sh}` is the sheath flow rate.

Example Usage

t,p = 295.15, 1e5                        
qsa,qsh = 1.66e-5, 3.33e-6                     
r₁,r₂,l = 9.37e-3,1.961e-2,0.44369               
Λ = DMAconfig(t,p,qsa,qsh,r₁,r₂,l,0.0,:-,6,:cylindrical) 
mobility = vtoz(Λ,1000.0) # [m2 V-1 s-1]

DifferentialMobilityAnalyzers.Ω — Method

Ω(Λ::DMAconfig, Z, zs)

The DMA transfer function is the probability that a particle of a particle of a given size exits the classifier via the sample flow. The diffusive broadened DMA transfer function is computed assuming blanced sheath and excess flows using the expression of Stolzenburg and McMurry (2008).

$\Omega(\tilde{z},\beta,\sigma) = \frac{\sigma}{\sqrt{2}\beta}\left[\epsilon \left( \frac{\tilde{z}-(1+\beta)}{\sqrt{2}\sigma} \right) + \epsilon \left (\frac{\tilde{z}- (1-\beta)}{\sqrt{2}\sigma} \right) - 2\epsilon \left ( \frac{\tilde{z}-1}{\sqrt{2}\sigma}\right) \right]$

where $\tilde{z} = \frac{z}{z^s}$ is the dimensionless mobility, $z$is the particle mobilityz^sis the centroid mobility selected by the DMA,\epsilon = x \mathrm{erf}(x) +\left(\exp(-x^2)/\sqrt{\pi}\right),\mathrm{erf}is the error function, and\beta = \frac{q_{sa}}{q_{sh}}. The parameter\sigmaaccounts for diffusional broading of the transfer function. Assuming plug flow,\sigma` can be computed using the following equations Hagwood (1999)

$\gamma = \left(\frac{r_1}{r_2}\right)^2$

$I = \frac{1}{2}(1+γ)$

$\kappa = \frac{lr_2}{r_2^2-r_1^2}$

$G = \frac{4(1+\beta)^2}{(1-γ)} \left[I+\{2(1+\beta)\kappa\}^{-2} \right ]$

$\sigma = \sqrt{\frac{2G\pi ld_{ab}}{q_{sh}}}$

Inputs for flow are taken from the DMAconfig. The function expects a mobility scalar z or vector Z, and a centroid mobility zˢ.

Example Usage

zˢ = dtoz(Λ, 200e-9)      # centroid mobility for Dp = 200 nm
z = [1e-9, 1e-8, 1e-7]    # mobility 
Ω(Λ,z,zˢ)                 # Output of the transfer function

Note

The function Ω is embedded in the the Type DifferentialMobilityAnalyzers.jl, which assigns δ.Ω either to this function Ω or Ωav applicable to scanning mode,

DifferentialMobilityAnalyzers.Ωav — Method

Ωav(Λ::DMAconfig, i::Int, k::Int; nint = 20)

Transfer function of the scanning DMA. The voltage continuously changes. Signal is acquired during some discrete time interval $t_c$. The SMPS transfer function is calculated as the average DMA transfer function during the time interval $[t,t+t_c]$ (Wang and Flagan, 1990).

$\Omega_{av} = \frac{1}{tc}\int_{t_i}^{t_i+t_c} \Omega(Z,z^s(t)) dt$

where $t_i$ is the start time when counting begins in channel $i$, $z^s(t)$ is the selected centroid mobility at time t and is calculated from the applied voltage.

Note

The function Ω is embedded in the the Type DifferentialMobilityAnalyzers.jl, which assigns δ.Ω either to this function Ωav or Ω applicable to stepping mode.

This function is used internally to compute the SMPS function and is tied to a specific DMA setup and scanning profile. It is called by the setupSMPS constructor functions. See setupSMPS in the source code to see how it is used.

DifferentialMobilityAnalyzers.getTc — Method

getTc(Λ::DMAconfig)

Returns a function

Tc(k::Integer, Dp)

to compute the charging efficiency. Tc depends of the polarity set in DMAconfig.

Charging efficiency (charge equilibrium) obtained in the bipolar charger is computed based on the parameterized measurements by Wiedensohler et al. (1988) with coefficients taken from the TSI 3080 Manual (2009).

$T_c(k) = 10^{\left\{ \sum_{i=1}^6 a_i (k) \left[ \ln \left(\frac{D_p}{nm}\right) \right]^{i-1} \right\}}$

where $k = -2,-1,1,2$ is the number and polarity of particle charge and $a_i$ are empirical coefficients.

For $k \ge \pm 3$, the formula from the TSI manual is used:

$T_c(k) = \frac{e}{\sqrt{4\pi^2\epsilon D_pk_bT}} \exp \left( \frac{-\frac{\left[|k| - 2\pi\epsilon D_pk_bT \ln(0.875)\right]^2}{e^2}}{ \frac{4\pi\epsilon D_pk_bT}{e^2}} \right)$

where $e$ is the elementary charge and $\epsilon$ is the dielectric constant for air.

Example Usage

t,p = 295.15, 1e5                        
qsa,qsh = 1.66e-5, 3.33e-6                     
r₁,r₂,l = 9.37e-3,1.961e-2,0.44369               
Λ = DMAconfig(t,p,qsa,qsh,r₁,r₂,l,0.0,:-,6,:cylindrical) 
Tc = getTc(Λ)
Tc(1,100.0) # Note that Dp is in units of nm!

Note

The function Tc is computed during DMA grid initialization and embedded in the DifferentialMobilityAnalyzer data type. It is usually accessed through this grid. The diameter Dp is in units of nm.

DifferentialMobilityAnalyzers.Tl — Method

Tl(Λ::DMAconfig, Dp)

Penetration efficiency through the TSI cylindrical DMA using the parameterization by Reineking & Porstendörfer (1986). The particle diameter Dp is in [nm].

$T_l = 0.82\exp(-11.5u)+0.1\exp(-70.0u)+0.03\exp(-180.0u)+0.02\exp(-340.0u)$

where $u = \frac{d_{ab} l_{eff}}{q_{sa}}$, $l_{eff}$ is the parameterized effective diffusion length, and $q_{sa}$ is the aerosol flow rate through the DMA.

Note

Λ contains the effective length, aerosol flow rate, temperature and pressure to compute $d_{ab}$. To treat multiple DMAs with different {leff, qsa, t, p} in a single script, the function Tl is embedded in the DifferentialMobilityAnalyzer data type.