Benchmarks

Functions to compute benchmarks.

DifferentialMobilityAnalyzers.benchmark — Method

benchmark(bins::Integer, num_threads::Integer)

Computes benchmarks for the three slowest operations, rinv, setupDMA, setupSMPS

bins is the number of DMA bins

num_threads is the number of BLAS threads

The function returns a dataframe that includes cpuinfo, juliaversion, blasvendor, blasthreads, number of bins, and the three timed benchmarks for rinv, setupDMA, setupSMPS.

DifferentialMobilityAnalyzers.runbenchmarks — Method

runbenchmarks()

Runs a set of standard benchmarks, varying the number bins and returns a dataframe with results. This function may take several minutes to complete.